requirements.txt
flask==3.1.*
yfinance>=0.2.40
numpy>=1.26
pandas>=2.0
gunicorn>=22.0
requests>=2.31
psycopg2-binary>=2.9
